Another large refactor to the XML Doc creation system and the module.
Moved the xml_data package into the xml_docs package; Moved all of the get_* functions from xml_builder.py to the new rst_utils.py and all of the get_<xml_class> functions to the __str__ and __repr__ functions of their respective XML classes. Updated the utility .bat scripts, and added a new script to clean the _build directory. Cython is no longer used, as it took too long to compile, and a local installation is preferable to a project-exclusive PyPi package. The src folder in the module has been renamed to egdocs, and egdocs is now loaded as the only package directly under the module.