Created egdocs.verification package
A beta version of the package was uploaded in a previous commit by mistake, but it is now ready to be tested for integration with egdocs.xml_docs. The egdocs.xml_docs.xml_data will also need to utilize verification, allowing full-category rebuilds if a template or an XML file is edited.