Updated Several Docs
Updated the index/readme to be more clear with several descriptions, also added sphinx.ext.todo to mark incomplete documentation. Updated the "data structures" document to include a warning about outdated information, as well as clarifying a few paragraphs. Changed the titles of file documents to include the name/purpose of the file in reasonable cases, with the extension now in parenthesis.